splitted the scorer into two files: file title - task description, file desc - task desc
